ShareWebNew/ShareWebUI/src/SuccessDialog/ui.desktop.tsx
import * as React from 'react';
import { noop } from 'lodash';
import UIIcon from '../UIIcon/ui.desktop'
import Panel from '../Panel/ui.desktop'
import Dialog from '../Dialog2/ui.desktop';
import __ from './locale';
import * as styles from './styles.client'

const SuccessDialog: React.StatelessComponent<UI.SuccessDialog.Props> = function SuccessDialog({
    onConfirm = noop,
    children
}) {
    return (
        <Dialog
            width={400}
            title={__('提示')}
            onClose={onConfirm}
            >
            <Panel>
                <Panel.Main>
                    <div className={styles['main']}>
                        <div className={styles['icon']}>
                            <UIIcon
                                code={'\uf0a2'}
                                color={'#54A67D'}
                                size={40}
                                />
                        </div>
                        <div className={styles['message']}>
                            {
                                children
                            }
                        </div>
                    </div>
                </Panel.Main>
                <Panel.Footer>
                    <Panel.Button
                        type="submit"
                        onClick={onConfirm}
                        >
                        {__('确定')}
                    </Panel.Button>
                </Panel.Footer>
            </Panel>
        </Dialog>
    )
}

export default SuccessDialog